requirements.txt
asgiref==3.2.3
Django==3.0
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
